Test of Gio-v against Gio-x
This is a big and ugly commit, containing new code for benchmarking, and a
few bug fixes.

The command for running benchmarks are:
go test -bench=. -cpu=1 -benchmem -cpuprofile profile.out

The program also prints som allocation and GC data to the console
when you change between the Gio-v and Gio-x windows.

It is set up to run as fast as possibl, reaching 120Hz on my PC.
The actual speed is shown in the first form.